fix(lint): corriger toutes les erreurs ESLint pour faire passer la CI
- eslint.config.js : restreindre l'analyse à src/** (exclut les dossiers Windows et website/)
- ThemeService.ts : any → unknown, catch(_){} → catch{}
- useSidebar.ts : catch(_){} → catch{}
- AssistantDrawer.tsx : déplacer le early return après les hooks (règle rules-of-hooks)
- PlanningSlotPicker.tsx : rendre toDateStr privée (fast-refresh)
- RecipeEditorShared.tsx : déplacer parsePrepTimeToMinutes/formatMinutes vers shared/utils/duration.ts (fast-refresh)
- SuggestionsPage.tsx : importer formatDate depuis shared/utils/date.ts
- RecipeDetailPage.tsx : any[] → MealieFavorite[], séparer le useEffect init/favorites
- RecipeDrawer.tsx : any[] → MealieFavorite[]
- RecipeFilters.tsx : any[] → MealieCategory[]/MealieTag[]
- AssistantService.ts : supprimer les eslint-disable devenus inutiles
