+++ src/presentation/components/PlanningSlotPicker.tsx
@@ -7,7 +7,7 @@ import { cn } from "../../lib/utils.ts"
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
 
-export function toDateStr(d: Date): string {
+function toDateStr(d: Date): string {
   const y = d.getFullYear()
   const m = String(d.getMonth() + 1).padStart(2, "0")
   const day = String(d.getDate()).padStart(2, "0")
