+++ src/presentation/components/RecipeEditorShared.tsx
@@ -5,34 +5,8 @@
 
 import { useState, useRef, useEffect, type ReactNode } from "react"
 import { Input } from "./ui/input.tsx"
-import { formatDuration } from "../../shared/utils/duration.ts"
+import { formatDuration, formatMinutes } from "../../shared/utils/duration.ts"
 import { cn } from "../../lib/utils.ts"
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-export function parsePrepTimeToMinutes(value?: string): string {
-  if (!value) return ""
-  if (/^\d+$/.test(value.trim())) {
-    const n = parseInt(value.trim(), 10)
-    return n > 0 ? String(n) : ""
-  }
-  const match = value.match(/PT(?:(\d+)H)?(?:(\d+)M)?/)
-  if (!match) return ""
-  const hours = parseInt(match[1] ?? "0")
-  const minutes = parseInt(match[2] ?? "0")
-  const total = hours * 60 + minutes
-  return total > 0 ? String(total) : ""
-}
-
-export function formatMinutes(value: string): string {
-  const n = Number(value)
-  if (!n || n <= 0) return ""
-  const h = Math.floor(n / 60)
-  const m = n % 60
-  if (h > 0 && m > 0) return `${h} h ${m} min`
-  if (h > 0) return `${h} h`
-  return `${m} min`
-}
 
 // ─── InlineEditText ───────────────────────────────────────────────────────────
